services/ml-anomaly-service/requirements.txt
flask
redis
scikit-learn
joblib
requests
pandas
numpy
python-jose[cryptography]